% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_utils.R
\name{as_dge}
\alias{as_dge}
\title{Coerce dense symmetric Matrix classes to general dense (dgeMatrix)}
\usage{
as_dge(A)
}
\arguments{
\item{A}{a dense Matrix}
}
\value{
a dgeMatrix
}
\description{
Replacement for the deprecated direct \code{as(., "dgeMatrix")}
coercion from dsyMatrix/dpoMatrix.
}
\keyword{internal}
