% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd_cache.R
\docType{data}
\name{.gmd_cache}
\alias{.gmd_cache}
\title{Internal cache for GMD factorizations}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.gmd_cache
}
\description{
Internal cache for GMD factorizations
}
\keyword{internal}
