% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd_cache.R
\docType{data}
\name{.gmd_cache_max_size}
\alias{.gmd_cache_max_size}
\title{Maximum number of cached entries (LRU eviction when exceeded)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.gmd_cache_max_size
}
\description{
Maximum number of cached entries (LRU eviction when exceeded)
}
\keyword{internal}
