% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd_cache.R
\docType{data}
\name{.gmd_cache_times}
\alias{.gmd_cache_times}
\title{Cache access times for LRU eviction}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.gmd_cache_times
}
\description{
Cache access times for LRU eviction
}
\keyword{internal}
