% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_utils.R
\name{.metric_rtol_default}
\alias{.metric_rtol_default}
\title{Default relative tolerance for metric validation}
\usage{
.metric_rtol_default()
}
\description{
Default relative tolerance for metric validation
}
\keyword{internal}
