% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpca-package.R
\docType{package}
\name{genpca-package}
\alias{genpca-package}
\title{genpca: Generalized Principal Component Analysis}
\description{
Generalized PCA and related matrix decompositions in weighted inner-product spaces. Methods are based on Allen, G. I., Grosenick, L., and Taylor, J. (2014) \doi{10.1080/01621459.2013.852978}, "A generalized least-square matrix decomposition", Journal of the American Statistical Association, 109(505), 145-159; and Abdi, H. (2007), "Singular value decomposition (SVD) and generalized singular value decomposition" \url{https://personal.utdallas.edu/~herve/Abdi-SVD2007-pretty.pdf}, in "Encyclopedia of Measurement and Statistics", 907-912.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/genpca/}
  \item \url{https://github.com/bbuchsbaum/genpca}
  \item Report bugs at \url{https://github.com/bbuchsbaum/genpca/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com} [copyright holder]

}
\keyword{internal}
