% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd_fast.R
\name{is_diagonal_metric}
\alias{is_diagonal_metric}
\title{Check whether a metric matrix is diagonal}
\usage{
is_diagonal_metric(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
logical
}
\description{
Check whether a metric matrix is diagonal
}
\keyword{internal}
