% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsutils.R
\name{is_identity_or_diag}
\alias{is_identity_or_diag}
\title{Check if constraint matrix is identity or purely diagonal with all != 1}
\usage{
is_identity_or_diag(M, eps = 1e-15)
}
\arguments{
\item{M}{A matrix (often \code{dsCMatrix}) or \code{NULL}.}

\item{eps}{Numeric tolerance}
}
\value{
TRUE if \code{M} is a (Matrix-based) diagonal with all or partial diag
}
\description{
Check if constraint matrix is identity or purely diagonal with all != 1
}
\keyword{internal}
