% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct data from a projection}
\usage{
reconstruct(x, comp, rowind, colind, ...)
}
\arguments{
\item{x}{A projection object}

\item{comp}{Components to use for reconstruction}

\item{rowind}{Row indices to reconstruct}

\item{colind}{Column indices to reconstruct}

\item{...}{Additional arguments}
}
\value{
Reconstructed data matrix
}
\description{
Re-exported from \pkg{multivarious}. See \code{\link[multivarious]{reconstruct}}
for details.
}
