% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_utils.R
\name{as_dgc}
\alias{as_dgc}
\title{Coerce to general CSC sparse matrix (dgCMatrix)}
\usage{
as_dgc(A)
}
\arguments{
\item{A}{a matrix or Matrix}
}
\value{
a dgCMatrix
}
\description{
Replacement for the deprecated direct `as(., "dgCMatrix")`
coercion from symmetric/triangular/diagonal Matrix classes.
}
\keyword{internal}
