% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{genpca}
\alias{genpca}
\title{Generalized Principal Components Analysis}
\usage{
genpca(
  X,
  A = NULL,
  M = NULL,
  ncomp = min(dim(X)),
  preproc = multivarious::center(),
  deflation = FALSE,
  threshold = 1e-06,
  use_cpp = TRUE,
  maxeig = 800
)
}
\arguments{
\item{X}{the data matrix}

\item{A}{the column constraints. Can be a \code{vector}, symmetric \code{matrix}, or symmetric sparse matrix with \code{ncol(X)} rows and columns.}

\item{M}{the row constraints. Can be a \code{vector}, symmetric \code{matrix}, or symmetric sparse matrix with \code{nrow(X)} rows and columns.}

\item{ncomp}{the number of components to return}

\item{preproc}{the type of pre-processing}

\item{deflation}{use the iterative deflation method}

\item{threshold}{convergence threshold for deflation method}

\item{use_cpp}{use the C++ implementation}
}
\description{
Compute a PCA in a inner-product space defined by row and coulmn constraint matrices.
}
\examples{
X <- matrix(rnorm(100*100), 100,100)
A <- cov(X)
M <- cov(t(X))
gp1 <- genpca(X, A=A, M=M, ncomp=100)

Xrecon <- reconstruct(gp1)
}
\references{
Abdi, H. (2007). Singular value decomposition (SVD) and generalized singular value decomposition. \emph{Encyclopedia of measurement and statistics}, 907-912.

Allen, G. I., Grosenick, L., & Taylor, J. (2014). A generalized least-square matrix decomposition. \emph{Journal of the American Statistical Association}, 109(505), 145-159.
}
