% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd_cache.R
\name{get_chol_lower_dense}
\alias{get_chol_lower_dense}
\title{Get (and cache) a *lower* Cholesky factor for a dense SPD matrix}
\usage{
get_chol_lower_dense(A)
}
\arguments{
\item{A}{numeric or dense Matrix (SPD). If sparse, falls back to dense.}
}
\value{
a base numeric matrix L (lower triangular) with A = L %*% t(L)
}
\description{
Get (and cache) a *lower* Cholesky factor for a dense SPD matrix
}
\keyword{internal}
